+++ docs/examples/single/controller.js
@@ -1,9 +1,9 @@
 angular.module('selectDemo')
     .controller('selectSingleController', function ($scope, ShopObj) {
 
-        $scope.shopObjShort = ShopObj.get();
+        $scope.shopObj = ShopObj.get();
 
-        $scope.shopObjShort.$promise.then(function(data) {
+        $scope.shopObj.$promise.then(function(data) {
             $scope.bundle = {a:1}; //data[3];
         })
     });